projects: eval-adis1647x: Update to new examples structure
Modify project to use the new examples structure implemented by
tools/scripts/examples.mk. Inspired from the eval-adxl355-pmdz project.
