+++ projects/eval-adis1647x/src/common/common_data.h
@@ -34,25 +34,12 @@
 #ifndef __COMMON_DATA_H__
 #define __COMMON_DATA_H__
 
-#include "platform_includes.h"
+#include "parameters.h"
 #include "adis1647x.h"
-#ifdef IIO_SUPPORT
-#include "iio_adis1647x.h"
-#ifdef IIO_TRIGGER_EXAMPLE
-#include "iio_trigger.h"
-#endif
-#endif
 
 extern struct no_os_uart_init_param adis1647x_uart_ip;
 extern struct no_os_spi_init_param adis1647x_spi_ip;
 extern struct no_os_gpio_init_param adis1647x_gpio_reset_ip;
 extern struct adis_init_param adis1647x_ip;
-
-#ifdef IIO_TRIGGER_EXAMPLE
-#define ADIS1647X_GPIO_TRIG_NAME "adis16477-1-dev0"
-
-extern struct iio_hw_trig_init_param adis1647x_gpio_trig_ip;
-extern struct no_os_irq_init_param adis1647x_gpio_irq_ip;
-#endif
 
 #endif /* __COMMON_DATA_H__ */
