+++ projects/eval-adis1647x/src/platform/maxim/main.c
@@ -31,17 +31,11 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
 
-#include "platform_includes.h"
+#include "parameters.h"
 #include "common_data.h"
 #include "no_os_error.h"
 
-#ifdef IIO_TRIGGER_EXAMPLE
-#include "iio_trigger_example.h"
-#endif
-
-#ifdef BASIC_EXAMPLE
-#include "basic_example.h"
-#endif
+extern int example_main();
 
 /**
  * @brief Main function execution for Maxim platform.
@@ -54,26 +48,10 @@ int main()
 
 	adis1647x_ip.spi_init = &adis1647x_spi_ip;
 
-#ifdef BASIC_EXAMPLE
-	struct no_os_uart_desc *uart_desc;
+#ifdef GPIO_DRDY_PORT_NUM
 
-	ret = no_os_uart_init(&uart_desc, &adis1647x_uart_ip);
-	if (ret)
-		return ret;
-
-	no_os_uart_stdio(uart_desc);
-	ret = basic_example_main();
-	no_os_uart_remove(uart_desc);
-#endif
-
-#ifdef IIO_TRIGGER_EXAMPLE
 	struct no_os_gpio_desc *adis_gpio_desc;
-	struct no_os_irq_ctrl_desc *nvic_desc;
-	struct no_os_irq_init_param nvic_ip = {
-		.platform_ops = &max_irq_ops,
-	};
 
-	/* Initialize DATA READY pin */
 	ret = no_os_gpio_get_optional(&adis_gpio_desc, &adis_gpio_drdy_ip);
 	if (ret)
 		return ret;
@@ -81,8 +59,14 @@ int main()
 	ret = no_os_gpio_direction_input(adis_gpio_desc);
 	if (ret)
 		goto remove_gpio;
+#endif
 
-	/* Initialize GPIO IRQ controller */
+#ifdef NVIC_GPIO_IRQ
+	struct no_os_irq_ctrl_desc *nvic_desc;
+	struct no_os_irq_init_param nvic_ip = {
+		.platform_ops = &max_irq_ops,
+	};
+
 	ret = no_os_irq_ctrl_init(&nvic_desc, &nvic_ip);
 	if (ret)
 		goto remove_gpio;
@@ -90,22 +74,14 @@ int main()
 	ret = no_os_irq_enable(nvic_desc, NVIC_GPIO_IRQ);
 	if (ret)
 		goto remove_irq_ctrl;
+#endif
 
-	ret = iio_trigger_example_main();
+	return example_main();
 
 remove_irq_ctrl:
 	no_os_irq_ctrl_remove(nvic_desc);
 remove_gpio:
 	no_os_gpio_remove(adis_gpio_desc);
-#endif
-
-#if (BASIC_EXAMPLE + IIO_TRIGGER_EXAMPLE == 0)
-#error At least one example has to be selected using y value in Makefile.
-#elif (BASIC_EXAMPLE + IIO_TRIGGER_EXAMPLE > 1)
-#error Selected example projects cannot be enabled at the same time. \
-Please enable only one example and rebuild the project.
-#endif
 
 	return ret;
 }
-
