+++ projects/eval-adis1647x/src/platform/pico/main.c
@@ -31,17 +31,11 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
 
-#include "platform_includes.h"
+#include "parameters.h"
 #include "common_data.h"
 #include "no_os_error.h"
 
-#ifdef IIO_TRIGGER_EXAMPLE
-#include "iio_trigger_example.h"
-#endif
-
-#ifdef BASIC_EXAMPLE
-#include "basic_example.h"
-#endif
+extern int example_main();
 
 /**
  * @brief Main function execution for pico platform.
@@ -53,27 +47,5 @@ int main()
 	int ret = -EINVAL;
 	adis1647x_ip.spi_init = &adis1647x_spi_ip;
 
-#ifdef BASIC_EXAMPLE
-	struct no_os_uart_desc *uart_desc;
-
-	ret = no_os_uart_init(&uart_desc, &adis1647x_uart_ip);
-	if (ret)
-		return ret;
-
-	ret = basic_example_main();
-	no_os_uart_remove(uart_desc);
-#endif
-
-#ifdef IIO_TRIGGER_EXAMPLE
-	ret = iio_trigger_example_main();
-#endif
-
-#if (BASIC_EXAMPLE + IIO_TRIGGER_EXAMPLE == 0)
-#error At least one example has to be selected using y value in Makefile.
-#elif (BASIC_EXAMPLE + IIO_TRIGGER_EXAMPLE > 1)
-#error Selected example projects cannot be enabled at the same time. \
-Please enable only one example and rebuild the project.
-#endif
-
-	return ret;
+	return example_main();
 }
